Improve greylisting permissions

+++ sigmapiweb/apps/PartyList/views.py
@@ -5,7 +5,12 @@ from django.contrib.auth.models import User
 from django.contrib.auth.decorators import permission_required, login_required
 from django.shortcuts import render, redirect
 
-from .models import Party, BlacklistedGuest, GreylistedGuest
+from .models import (
+    BlacklistedGuest,
+    GreylistedGuest,
+    Party,
+    user_can_delete_greylisting,
+)
 from .forms import PartyForm, EditPartyInfoForm, BlacklistForm, GreylistForm
 
 
@@ -76,10 +81,7 @@ def manage_blacklist(request):
     """
     View for managing blacklist.
     """
-    context = {
-        'blacklist': BlacklistedGuest.objects.all(),
-        'message': None
-    }
+    message = None
     if request.method == 'POST':
         form = BlacklistForm(request.POST)
         if form.is_valid():
@@ -88,27 +90,28 @@ def manage_blacklist(request):
             if not form.cleaned_data['details']:
                 new_blacklisted_guest.details = ''
             new_blacklisted_guest.save()
-            context['message'] = 'Successfully added entry to blacklist'
+            message = 'Successfully added entry to blacklist'
         else:
-            context['message'] = 'Error adding entry to blacklist'
+            message = 'Error adding entry to blacklist'
     else:
         form = BlacklistForm()
-    context['form'] = form
+    context = {
+        'blacklist': BlacklistedGuest.objects.all(),
+        'message': message,
+        'form': form,
+    }
     return render(request, 'parties/blacklist/manage.html', context)
 
 
 @permission_required(
-    'PartyList.manage_greylist',
+    'PartyList.add_greylistedguest',
     login_url='pub-permission_denied',
 )
 def manage_greylist(request):
     """
     View for managing greylist.
     """
-    context = {
-        'greylist': GreylistedGuest.objects.all(),
-        'message': None
-    }
+    message = None
     if request.method == 'POST':
         form = GreylistForm(request.POST)
         if form.is_valid():
@@ -116,13 +119,24 @@ def manage_greylist(request):
             new_greylisted_guest = form.save(commit=False)
             if not form.cleaned_data['details']:
                 new_greylisted_guest.details = ''
+            new_greylisted_guest.greylister = request.user
             new_greylisted_guest.save()
-            context['message'] = 'Successfully added entry to greylist'
+            message = 'Successfully added entry to greylist'
         else:
-            context['message'] = 'Error adding entry to greylist'
+            message = 'Error adding entry to greylist'
     else:
         form = GreylistForm()
-    context['form'] = form
+    context = {
+        'greylist': [
+            (
+                greylisting,
+                user_can_delete_greylisting(request.user, greylisting),
+            )
+            for greylisting in GreylistedGuest.objects.all().order_by('name')
+        ],
+        'message': message,
+        'form': form,
+    }
     return render(request, 'parties/greylist/manage.html', context)
 
 
@@ -141,16 +155,19 @@ def remove_blacklisting(request, bl_id):
 
 
 @permission_required(
-    'PartyList.manage_greylist',
+    'PartyList.delete_greylistedguest',
     login_url='pub-permission_denied',
 )
-def remove_greylisting(request, bl_id):
+def remove_greylisting(request, gl_id):
     """
     View for removing a greylisted guest.
     """
     if request.method == 'POST':
-        gl_guest = GreylistedGuest.objects.get(pk=bl_id)
-        gl_guest.delete()
+        gl_guest = GreylistedGuest.objects.get(pk=gl_id)
+        if user_can_delete_greylisting(request.user, gl_guest):
+            gl_guest.delete()
+        else:
+            return redirect('pub-permission_denied')
     return redirect('partylist-manage_greylist')
